New changes to fix issues brought up by array

+++ src/boot/ksdk1.1.0/devMMA8451Q.h
@@ -41,4 +41,4 @@ WarpStatus	readSensorRegisterMMA8451Q(uint8_t deviceRegister, int numberOfBytes)
 WarpStatus	writeSensorRegisterMMA8451Q(uint8_t deviceRegister,
 					uint8_t payloadBtye);
 WarpStatus	configureSensorMMA8451Q(uint8_t payloadF_SETUP, uint8_t payloadCTRL_REG1);
-uint16_t		printSensorDataMMA8451Q(bool hexModeFlag, uint16_t store[]);
+uint16_t		printSensorDataMMA8451Q(bool hexModeFlag, int store[]);
